pyproject.toml
[tool.poetry]
name = "napalm-core"
version = "0.1.0"
description = "Your project to manage napalm detection modules"
license = "AGPL-3.0"
readme = "README.md"
authors = []
packages = [
    { include = "napalm_core" },
]

[tool.poetry.dependencies]
python = "^3.10"
slither-analyzer = "^0.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.plugins."napalm.collection"]
"napalm-base" = "napalm_base.napalm:entry_point"

